+++ app/src/main/java/com/example/proyecto_integrador/Order_ticket.kt
@@ -1,10 +1,10 @@
 package com.example.proyecto_integrador
 
-data class Order_ticket(val title: String, val quantity: Int, val price: Float, val Stotal: Double,val image : Int ){
+data class Order_ticket(val title: String, val quantity: Int, val price: Double, val Stotal: Double,val image : Int ){
 
     private var _title : String? = null
     private var _quantity : Int? = null
-    private var _price : Float? = null
+    private var _price : Double? = null
     private var _Stotal : Double? = null
     private var _image : Int? = null
